Add update mechanicsm with concurrency check so we should haven't duplicate feeds but added a hash to the rss items table just in case.
Add Simple update mechanism to feed viewer to will update the database and
reload the page.

Add favicon.ico to match project logo
