+++ src/FeedReader/FeedReader.csproj
@@ -181,6 +181,7 @@
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\RssModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Providers\RssProvider.cs" />
     <Compile Include="Providers\RssSearchProvider.cs" />
     <Compile Include="Providers\RssUpdater.cs" />
     <Compile Include="Providers\SyndicationFeedXmlReader.cs" />
